Hot Fix

+++ src/task1.cpp
@@ -8,7 +8,7 @@ int comp1(const void* a, const void* b)
 
 int comp2(const void* a, const void* b)
 {
-	return (*(double*)a - *(double*)b);
+	return (int)(*(double*)a - *(double*)b);
 }
 
 int comp3(const void* a, const void* b)
